[CDEC-378] Banish more orphan Bi instances from core
