+++ core/src/Pos/Aeson/Genesis.hs
@@ -44,7 +44,6 @@ import           Formatting (sformat)
 import           Serokell.Aeson.Options (defaultOptions)
 
 import           Pos.Aeson.Core ()
-import           Pos.Binary.Core.Address ()
 import           Pos.Core.Common (Address, Coin, StakeholderId, unsafeGetCoin)
 import           Pos.Core.Delegation (ProxySKHeavy)
 import           Pos.Core.Genesis (FakeAvvmOptions, GenesisAvvmBalances (..), GenesisDelegation,
