--- core/src/Pos/Binary/Core.hs
@@ -1,10 +1,0 @@
-{-# OPTIONS_GHC -Wno-unused-imports   #-}
-{-# OPTIONS_GHC -Wno-dodgy-exports    #-}
-
-module Pos.Binary.Core
-       ( module Pos.Binary.Core.Address
-       , module Pos.Binary.Core.Script
-       ) where
-
-import           Pos.Binary.Core.Address
-import           Pos.Binary.Core.Script ()
