+++ core/src/Pos/Core/Binary.hs
@@ -1,0 +1,8 @@
+{-# OPTIONS_GHC -Wno-unused-imports   #-}
+{-# OPTIONS_GHC -Wno-dodgy-exports    #-}
+
+module Pos.Core.Binary
+       ( module Pos.Core.Binary.Orphans
+       ) where
+
+import           Pos.Core.Binary.Orphans ()
