+++ core/src/Pos/Core/Common/Address.hs
@@ -17,6 +17,10 @@ module Pos.Core.Common.Address
        , checkScriptAddress
        , checkRedeemAddress
 
+       -- * Encoding
+       , encodeAddr
+       , encodeAddrCRC32
+
        -- * Utilities
        , addrAttributesUnwrapped
        , deriveLvl2KeyPair
@@ -60,7 +64,7 @@ import qualified Data.Text.Buildable as Buildable
 import           Formatting (Format, bprint, build, builder, later, (%))
 import           Serokell.Data.Memory.Units (Byte)
 
-import           Pos.Binary.Class (Bi, biSize)
+import           Pos.Binary.Class (Bi (..), Encoding, biSize, encodeCrcProtected)
 import qualified Pos.Binary.Class as Bi
 import           Pos.Core.Common.Coin ()
 import           Pos.Core.Constants (accountGenesisIndex, wAddressGenesisIndex)
@@ -415,3 +419,16 @@ goodPk = fst goodSkAndPk
 
 goodSk :: SecretKey
 goodSk = snd goodSkAndPk
+
+
+-- Encodes the `Address` __without__ the CRC32.
+-- It's important to keep this function separated from the `encode`
+-- definition to avoid that `encode` would call `crc32` and
+-- the latter invoke `crc32Update`, which would then try to call `encode`
+-- indirectly once again, in an infinite loop.
+encodeAddr :: Address -> Encoding
+encodeAddr Address {..} =
+    encode addrRoot <> encode addrAttributes <> encode addrType
+
+encodeAddrCRC32 :: Address -> Encoding
+encodeAddrCRC32 Address{..} = encodeCrcProtected (addrRoot, addrAttributes, addrType)
