+++ core/src/Pos/Core/Ssc/Commitment.hs
@@ -12,8 +12,10 @@ import qualified Data.HashMap.Strict as HM
 
 import           Pos.Core.Slotting (EpochIndex)
 
-import           Pos.Binary.Class (AsBinary, fromBinary, serialize')
+import           Pos.Binary.Class (AsBinary, Bi (..), encodeListLen, enforceSize, fromBinary,
+                                   serialize')
 import           Pos.Crypto (EncShare, PublicKey, SecretProof, Signature, VssPublicKey)
+import           Pos.Util.Util (cborError)
 
 -- | Commitment is a message generated during the first stage of SSC.
 -- It contains encrypted shares and proof of secret.
@@ -31,6 +33,17 @@ instance Hashable Commitment
 instance Ord Commitment where
     compare = comparing (serialize' . commProof) <>
               comparing (sort . HM.toList . commShares)
+
+instance Bi Commitment where
+    encode Commitment{..} = encodeListLen 2 <> encode commShares
+                                            <> encode commProof
+    decode = do
+        enforceSize "Commitment" 2
+        commShares <- decode
+        when (null commShares) $ cborError "decode@Commitment: no shares"
+        commProof <- decode
+        return $ Commitment commProof commShares
+
 
 -- | Signature which ensures that commitment was generated by node
 -- with given public key for given epoch.
