+++ core/src/Pos/Core/Update/ApplicationName.hs
@@ -10,9 +10,15 @@ import           Control.Monad.Except (MonadError (throwError))
 import           Data.Char (isAscii)
 import qualified Data.Text as T
 
+import           Pos.Binary.Class (Bi (..))
+
 newtype ApplicationName = ApplicationName
     { getApplicationName :: Text
     } deriving (Eq, Ord, Show, Generic, Typeable, ToString, Hashable, Buildable, NFData)
+
+instance Bi ApplicationName where
+    encode appName = encode (getApplicationName appName)
+    decode = ApplicationName <$> decode
 
 -- | Smart constructor of 'ApplicationName'.
 checkApplicationName :: MonadError Text m => ApplicationName -> m ()
@@ -25,3 +31,4 @@ checkApplicationName (ApplicationName appName)
 
 applicationNameMaxLength :: Integral i => i
 applicationNameMaxLength = 12
+
