+++ core/src/Pos/Core/Update/BlockVersion.hs
@@ -9,6 +9,7 @@ import qualified Data.Text.Buildable as Buildable
 import           Formatting (bprint, shown)
 import qualified Prelude
 
+import           Pos.Binary.Class (Cons (..), Field (..), deriveSimpleBi)
 import           Pos.Util.Some (Some, liftLensSome)
 
 -- | Communication protocol version.
@@ -34,3 +35,10 @@ class HasBlockVersion a where
 
 instance HasBlockVersion (Some HasBlockVersion) where
     blockVersionL = liftLensSome blockVersionL
+
+deriveSimpleBi ''BlockVersion [
+    Cons 'BlockVersion [
+        Field [| bvMajor :: Word16 |],
+        Field [| bvMinor :: Word16 |],
+        Field [| bvAlt   :: Word8  |]
+    ]]
