+++ core/src/Pos/Core/Update/Proof.hs
@@ -3,7 +3,6 @@ module Pos.Core.Update.Proof
        , mkUpdateProof
        ) where
 
-import           Pos.Binary.Class (Bi)
 import           Pos.Crypto (Hash, hash)
 
 import           Pos.Core.Update.Payload
@@ -11,7 +10,5 @@ import           Pos.Core.Update.Payload
 -- | Proof that body of update message contains 'UpdatePayload'.
 type UpdateProof = Hash UpdatePayload
 
-mkUpdateProof
-    :: Bi UpdatePayload
-    => UpdatePayload -> UpdateProof
+mkUpdateProof :: UpdatePayload -> UpdateProof
 mkUpdateProof = hash
