+++ core/src/Pos/Core/Update/SystemTag.hs
@@ -19,11 +19,17 @@ import           Distribution.Text (display)
 import           Instances.TH.Lift ()
 import           Language.Haskell.TH.Syntax (Lift)
 
+import           Pos.Binary.Class (Bi (..))
+
 -- | Tag of system for which update data is purposed, e.g. win64, mac32
 newtype SystemTag = SystemTag { getSystemTag :: Text }
   deriving (Eq, Ord, Show, Generic, Buildable, Hashable, Lift, Typeable)
 
 instance NFData SystemTag
+
+instance Bi SystemTag where
+    encode = encode . getSystemTag
+    decode = SystemTag <$> decode
 
 systemTagMaxLength :: Integral i => i
 systemTagMaxLength = 10
