+++ core/test/Test/Pos/Core/CborSpec.hs
@@ -17,7 +17,6 @@ import           Test.QuickCheck (Arbitrary (..))
 import           Test.QuickCheck.Arbitrary.Generic (genericArbitrary, genericShrink)
 
 import           Pos.Binary.Class (Bi (..), Cons (..), Field (..), deriveSimpleBi, serialize, unsafeDeserialize)
-import           Pos.Binary.Core ()
 import           Pos.Core
 import           Pos.Core.Common (ScriptVersion)
